requirements.txt
pyautogui~=0.9.54
pydirectinput~=1.0.4
keyboard~=0.13.5
pillow~=10.4.0
stopwatch.py~=2.0.1
opencv-python~=4.10.0.84
PyScreeze~=1.0.1
pywin32
numpy